Run `import`s of all base modules before any extension
This ensures that all base modules are `import`ed in the exact same
order for every extension.

This also for some reason is required to make firebase listeners work.
react-native-camera has something that triggers something when `import`ed,
which causes firebase listeners to work.
I suspect `requireNativeComponent`. I'm not sure what exactly the
react native impl of firebase is needing to be initialized that is taken
care of by `import`ing react-native-camera.
For this reason, and general consistency, we import all the base modules
in the same order at the beginning every time.